+++ KeyboardKitDemoKeyboard/KeyboardViewController+Keyboard.swift
@@ -34,7 +34,6 @@ extension KeyboardViewController {
         let keyboard = EnhancedEmojiKeyboard(in: self)
         let config = keyboard.gridConfig
         let view = HFloatingHeaderButtonCollectionView(id: "EnhancedEmojiKeyboard", categoryActions: keyboard.categoryActions, configuration: config, buttonCreator: { [unowned self] in return self.button(for: $0) })
-        
         let bottom = buttonRow(for: keyboard.bottomActions, distribution: .fillProportionally)
         keyboardStackView.addArrangedSubview(view)
         keyboardStackView.addArrangedSubview(bottom)