+++ KeyboardKitDemoKeyboard_SwiftUI/KeyboardView+Keyboards.swift
@@ -20,7 +20,10 @@ extension KeyboardView {
     
     var systemKeyboard: some View {
         VStack(spacing: 0) {
-            AutocompleteToolbar(suggestions: autocompleteContext.suggestions, buttonBuilder: autocompleteButtonBuilder).frame(height: 50)
+            AutocompleteToolbar(
+                suggestions: autocompleteContext.suggestions,
+                buttonBuilder: autocompleteButtonBuilder)
+                .frame(height: 50)
             KeyboardButtonWidthPreview()
             SystemKeyboard(
                 layout: systemKeyboardLayout,