+++ Sources/KeyboardKit/Layout/View+KeyboardLayoutWidth.swift
@@ -1,5 +1,5 @@
 //
-//  View+KeyboardButtonWidth.swift
+//  View+KeyboardLayoutWidth.swift
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2021-01-21.
@@ -20,7 +20,7 @@ public extension View {
      using `.reference` and `.fromReference` width types.
      */
     @ViewBuilder
-    func width(_ width: KeyboardButtonWidth, totalWidth: CGFloat = .zero, referenceSize: Binding<CGSize> = .constant(.zero)) -> some View {
+    func width(_ width: KeyboardLayoutWidth, totalWidth: CGFloat = .zero, referenceSize: Binding<CGSize> = .constant(.zero)) -> some View {
         switch width {
         case .fromReference: self.frame(width: referenceSize.width.wrappedValue)
         case .percentage(let percent): self.frame(width: percent * totalWidth)
@@ -36,7 +36,7 @@ private extension View {
     /**
      Calculate the width in points for a certain width type.
      */
-    func width(for width: KeyboardButtonWidth, totalWidth: CGFloat, referenceSize: CGSize) -> CGFloat? {
+    func width(for width: KeyboardLayoutWidth, totalWidth: CGFloat, referenceSize: CGSize) -> CGFloat? {
         switch width {
         case .percentage(let percent): return percent * totalWidth
         case .points(let points): return points